% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{get_discriminating_features}
\alias{get_discriminating_features}
\title{Helper function to extract the discriminating features}
\usage{
get_discriminating_features(
  counts1,
  counts2,
  label1,
  label2,
  de_res = NULL,
  verbosity = 0
)
}
\arguments{
\item{counts1}{Count matrix for group 1 (columns are cells)}

\item{counts2}{Count matrix for group 2 (columns are cells)}

\item{label1}{The label of group 1}

\item{label2}{The label of group 2}

\item{de_res}{Optional; results of differential expression tests
(output of sctransform::diff_mean_test with compare argument set to 'all_vs_all');
if not provided, DE testing will be done during training; default is NULL}

\item{verbosity}{Level of printed output; 0 is quiet; default is 0}
}
\description{
Helper function to extract the discriminating features
}
