% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pg_model_human_lung}
\alias{pg_model_human_lung}
\title{Pre-trained model for human lung}
\format{
A list - the output of Perigee::pg_train
}
\source{
\url{https://azimuth.hubmapconsortium.org/references/#Human\%20-\%20Lung}
}
\usage{
pg_model_human_lung
}
\description{
Model trained on human lung and circulating blood. 39 cell types.
See https://azimuth.hubmapconsortium.org/references/#Human\%20-\%20Lung (annotation level 1).
}
\keyword{datasets}
