% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pg_model_human_pbmc}
\alias{pg_model_human_pbmc}
\title{Pre-trained model for human PBMC}
\format{
A list - the output of Perigee::pg_train
}
\source{
\url{https://azimuth.hubmapconsortium.org/references/#Human\%20-\%20PBMC}
}
\usage{
pg_model_human_pbmc
}
\description{
Model trained on Human PBMC. 29 non-doublet cell types.
See https://azimuth.hubmapconsortium.org/references/#Human\%20-\%20PBMC (annotation level 2).
}
\keyword{datasets}
