% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{pg_predict}
\alias{pg_predict}
\title{Cell type prediction with Perigee}
\usage{
pg_predict(pg_model, counts, return_dcs = FALSE, verbosity = 1)
}
\arguments{
\item{pg_model}{A Perigee model created with pg_train}

\item{counts}{Matrix of single-cell RNA-seq count data}

\item{return_dcs}{Should the output contain the full matrix of pairwise
differences of cosine similarities? Default is FALSE}

\item{verbosity}{0 is quiet; larger integers may lead to more print
output; default is 1}
}
\value{
A data frame of predictions with columns

Output columns:
\describe{
\item{label}{The predicted cell type label}
\item{n_comp}{The number of pairwise comparisons that support the predicted label}
\item{score}{The prediction score (based on model specificity)}
}
}
\description{
Cell type prediction with Perigee
}
