% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{pg_train}
\alias{pg_train}
\title{Training a Perigee model}
\usage{
pg_train(
  counts,
  cell_labels,
  de_res = NULL,
  max_cells = NULL,
  exclude = NULL,
  verbosity = 1
)
}
\arguments{
\item{counts}{Matrix of single-cell RNA-seq count data; columns are cells}

\item{cell_labels}{Cell type labels}

\item{de_res}{Optional; results of differential expression tests
(output of sctransform::diff_mean_test with compare argument set to 'all_vs_all');
if NULL, DE testing will be done during training; default is NULL}

\item{max_cells}{Maximum number of cells to use per type; default uses all cells}

\item{exclude}{Cell type labels to exclude from training and the final model}

\item{verbosity}{Level of printed output; 0 is quiet; 1 provides progress information;
2 provide more details; default is 1}
}
\value{
A Perigee model (named list)
}
\description{
Training a Perigee model
}
